made email longer, fixed bug with form not getting object_data, made the Exists() validator generic, moved it to 'lib' folder, applied it to name_in_util slightly edited alembic config to detect column type changes
