+++ project/auth/validators.py
@@ -30,25 +30,3 @@ class PasswordCorrect(object):
             raise ValidationError(self.message)
 
 
-class Exists(object):
-    def __init__(self, message=None, reverse=False):
-        if message:
-            self.message = message
-        self.reverse = reverse
-
-    def __call__(self, _, field):
-        if field.object_data == field.data:
-            return
-        if not hasattr(self, 'message'):
-            if not self.reverse:
-                self.message = 'Пользователь с таким {} уже существует'\
-                    .format(field.name)
-            else:
-                self.message = 'Пользователь с таким {} не существует'\
-                    .format(field.name)
-
-        u = True \
-            if list(User.query.filter(getattr(User, field.name) == field.data)) \
-            else False
-        if self.reverse ^ u:
-                raise ValidationError(self.message)
